Adjust loop iterations and outputs

+++ EnergyTest/Benchmarks/C/PolynomialEvaluation.h
@@ -6,6 +6,6 @@
 static double x = 5.5;
 
 void InitCS();
-double HornersRule(long LoopIterations);
+long double HornersRule(long LoopIterations);
 
 #endif //C_POLYNOMIALEVALUATION_H
